Hierarchical Injectors
set most popular book is in data service which is used by dashboard and edit book.but if edit book has dataservice provider in local, it won't look to the root injector.so set popular book in edit book page won't rflect in dash board.

component look for providers in local injector>parent injector>root injector.
this is the order.
if same service is in local and root providers then 2 separeate instances will be created like in this example.